+++ packages/@n8n/agents/src/__tests__/integration/agent-vector-store-qdrant.test.ts
@@ -5,11 +5,14 @@
  * keys; self-skips in CI replay since Qdrant can't be replayed from cassettes.
  */
 import type { QdrantClient } from '@qdrant/js-client-rest';
-import { afterAll, beforeAll, describe, expect, it } from 'vitest';
+import { afterAll, beforeAll, describe } from 'vitest';
 
-import { findLastTextContent } from './helpers';
-import { loadBulkFixture } from './vector-store-helpers';
-import { Agent, VectorStore } from '../../index';
+import {
+	loadBulkFixture,
+	registerAgentVectorStoreTests,
+	upsertFixtureDocuments,
+} from './vector-store-helpers';
+import { VectorStore } from '../../index';
 import { QdrantVectorStore } from '../../vector-stores/qdrant';
 
 const QDRANT_URL = process.env.QDRANT_TEST_URL;
@@ -41,14 +44,7 @@ describe.skipIf(!QDRANT_URL || !hasKeys)('Agent + QdrantVectorStore end-to-end',
 			apiKey: QDRANT_API_KEY,
 			collectionName,
 		});
-		await store.upsert(
-			fixture.documents.map((doc) => ({
-				id: doc.id,
-				vector: doc.vector,
-				content: doc.content,
-				metadata: doc.metadata,
-			})),
-		);
+		await upsertFixtureDocuments(store, fixture);
 		knowledge = new VectorStore('knowledge-base')
 			.store(store)
 			.embeddingModel('openai/text-embedding-3-small')
@@ -60,74 +56,5 @@ describe.skipIf(!QDRANT_URL || !hasKeys)('Agent + QdrantVectorStore end-to-end',
 		await store.close();
 	});
 
-	it('answers a question using knowledge retrieved from Qdrant', async () => {
-		const agent = new Agent('kb-assistant')
-			.model('anthropic/claude-haiku-4-5')
-			.instructions(
-				'You answer questions from the knowledge base. Always search it before answering. Be concise.',
-			)
-			.vectorStore(knowledge);
-
-		const result = await agent.generate(
-			'What is the tallest federal building in Manhattan, and how many stories does it have?',
-		);
-
-		const searchCalls = (result.toolCalls ?? []).filter(
-			(tc) => tc.tool === 'search_knowledge_base',
-		);
-		expect(searchCalls.length).toBeGreaterThanOrEqual(1);
-
-		const searchOutput = searchCalls[0].output as {
-			results: Array<{ content: string; score: number; metadata: { title?: string } }>;
-		};
-		expect(searchOutput.results.length).toBeGreaterThanOrEqual(1);
-		expect(searchOutput.results[0].metadata.title).toBe('Jacob K. Javits Federal Building');
-
-		const answer = findLastTextContent(result.messages);
-		expect(answer).toMatch(/41/);
-	});
-
-	it('narrows results with a model-controlled metadata filter', async () => {
-		const agent = new Agent('kb-assistant-filtered')
-			.model('anthropic/claude-haiku-4-5')
-			.instructions(
-				'You answer questions from the knowledge base. Always search it before answering. ' +
-					'Always pass a filter on the category key when the user names a specific category.',
-			)
-			.vectorStore(knowledge, {
-				filterableKeys: {
-					category: "Document category, exactly one of: 'history', 'science', 'geography'",
-				},
-			});
-
-		const result = await agent.generate(
-			'Search only the geography category: what places are described?',
-		);
-
-		const searchCalls = (result.toolCalls ?? []).filter(
-			(tc) => tc.tool === 'search_knowledge_base',
-		);
-		expect(searchCalls.length).toBeGreaterThanOrEqual(1);
-
-		const input = searchCalls[0].input as {
-			filter?: Array<{ key: string; operator: string; value: unknown }>;
-		};
-		const categoryCondition = input.filter?.find((c) => c.key === 'category');
-		expect(categoryCondition).toBeDefined();
-		if (categoryCondition?.operator === 'in') {
-			expect(categoryCondition.value).toContain('geography');
-		} else {
-			expect(categoryCondition?.value).toBe('geography');
-		}
-
-		const searchOutput = searchCalls[0].output as {
-			results: Array<{ metadata: { category?: string } }>;
-		};
-		for (const searchResult of searchOutput.results) {
-			expect(searchResult.metadata.category).toBe('geography');
-		}
-
-		const answer = findLastTextContent(result.messages);
-		expect(answer).toBeTruthy();
-	});
+	registerAgentVectorStoreTests(() => knowledge);
 });
