+++ packages/@n8n/agents/src/__tests__/integration/vector-store-bulk-supabase.test.ts
@@ -7,13 +7,14 @@
  * backend actually returns.
  */
 import type { Pool } from 'pg';
-import { afterAll, beforeAll, describe, expect, it } from 'vitest';
+import { afterAll, beforeAll, describe } from 'vitest';
 
 import {
 	createSupabaseVectorTableAndFunction,
 	dropSupabaseVectorTableAndFunction,
 	loadBulkFixture,
-	localTopIds,
+	registerBulkVectorStoreTests,
+	upsertFixtureDocuments,
 	waitUntilQueryable,
 } from './vector-store-helpers';
 import { SupabaseVectorStore } from '../../vector-stores/supabase';
@@ -46,14 +47,7 @@ describe.skipIf(!SUPABASE_URL || !SUPABASE_API_KEY || !SUPABASE_DB_URL)(
 				queryName,
 			});
 			await waitUntilQueryable(store, fixture.dimensions);
-			await store.upsert(
-				fixture.documents.map((doc) => ({
-					id: doc.id,
-					vector: doc.vector,
-					content: doc.content,
-					metadata: doc.metadata,
-				})),
-			);
+			await upsertFixtureDocuments(store, fixture);
 		}, HOOK_TIMEOUT_MS);
 
 		afterAll(async () => {
@@ -62,64 +56,6 @@ describe.skipIf(!SUPABASE_URL || !SUPABASE_API_KEY || !SUPABASE_DB_URL)(
 			await pool.end();
 		});
 
-		it('returns the expected best match for each known-answer query', async () => {
-			for (const query of fixture.queries) {
-				const results = await store.query(query.vector, { topK: 5 });
-
-				expect(results).toHaveLength(5);
-				expect(results[0].id).toBe(query.expectedTopId);
-				expect(results[0].content.length).toBeGreaterThan(0);
-				expect(typeof results[0].metadata.title).toBe('string');
-				for (let i = 1; i < results.length; i++) {
-					expect(results[i].score).toBeLessThanOrEqual(results[i - 1].score);
-				}
-			}
-		});
-
-		it('filters by category over the large corpus and matches locally computed ground truth', async () => {
-			const scienceDocs = fixture.documents.filter((doc) => doc.metadata.category === 'science');
-			const expectedIds = localTopIds(scienceDocs, fixture.queries[0].vector, 10);
-
-			const results = await store.query(fixture.queries[0].vector, {
-				topK: 10,
-				filter: { conditions: [{ key: 'category', operator: 'eq', value: 'science' }] },
-			});
-
-			expect(results.map((r) => r.id)).toEqual(expectedIds);
-			for (const result of results) expect(result.metadata.category).toBe('science');
-		});
-
-		it('nin filter excludes categories across the corpus', async () => {
-			const geographyDocs = fixture.documents.filter(
-				(doc) => doc.metadata.category === 'geography',
-			);
-			const expectedIds = localTopIds(geographyDocs, fixture.queries[0].vector, 10);
-
-			const results = await store.query(fixture.queries[0].vector, {
-				topK: 10,
-				filter: {
-					conditions: [{ key: 'category', operator: 'nin', value: ['history', 'science'] }],
-				},
-			});
-
-			expect(results.map((r) => r.id)).toEqual(expectedIds);
-			for (const result of results) expect(result.metadata.category).toBe('geography');
-		});
-
-		it('deletes a batch and stops returning the deleted documents', async () => {
-			const idsToDelete = localTopIds(fixture.documents, fixture.queries[1].vector, 20);
-			expect(idsToDelete).toContain(fixture.queries[1].expectedTopId);
-
-			await store.delete({ ids: idsToDelete });
-
-			const remaining = fixture.documents.filter((doc) => !idsToDelete.includes(doc.id));
-			const expectedTopIds = localTopIds(remaining, fixture.queries[1].vector, 5);
-
-			const results = await store.query(fixture.queries[1].vector, { topK: 5 });
-			const resultIds = results.map((r) => r.id);
-
-			expect(resultIds).toEqual(expectedTopIds);
-			for (const deletedId of idsToDelete) expect(resultIds).not.toContain(deletedId);
-		});
+		registerBulkVectorStoreTests(fixture, () => store);
 	},
 );
